src/main/java/dev/mgbarbosa/urlshortner/services/SecurityServiceImpl.java
package dev.mgbarbosa.urlshortner.services;

import dev.mgbarbosa.urlshortner.services.interfaces.SecurityService;
import org.mindrot.jbcrypt.BCrypt;
import org.springframework.stereotype.Service;

@Service
public class SecurityServiceImpl implements SecurityService {
    @Override
    public String hashPassword(String input) {
        return BCrypt.hashpw(input, BCrypt.gensalt());
    }

    @Override
    public boolean verifyPassword(String password, String hash) {
        return BCrypt.checkpw(password, hash);
    }
}
